motor-finder-python/requirements.txt
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
bidict==0.23.1
boto3==1.38.34
botocore==1.38.34
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.1
orjson==3.10.18
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-engineio==4.12.2
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
rsa==4.9.1
s3transfer==0.13.0
sendgrid==6.12.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
